Activity-driven network model of Cai, Nie & Holme (2024).

Here, nodes are initially inactive and have degree zero. Node \(i\) activates with rate \(a[i] * \eta\) and upon activation connects to \(m\) other uniformly chosen nodes (which are not necessarily active). Active nodes inactivate with constant rate \(b\). See Cai, Nie & Holme 2024, Phys. Rev. Research 6, L022017 for details. Here, we implement a generalized version of the model in which the activation and deacivation rates of infected nodes can differ from those of non-infected node.

Usage

activity_driven_temporalnetwork(
  activities,
  m,
  eta,
  b,
  eta_inf = eta,
  b_inf = b
)

Arguments

activities

node-specific activities \(a_1,a_2,\ldots\)

m

number of nodes an activated node connects to

eta

activation rate

b

deactivation rate

eta_inf

activation rate for infected nodes

b_inf

deactivation rate of infected nodes

See also