Create a Brownian proximity network

Create a Brownian proximity network. size nodes are places randomly in two dimensions, and connected by links if their distance does not exceed radius. The size of the playing field is chosen such that the expected number of neighbours of each node is avg_degree.

The resulting network is temporal, i.e. evolves over time. Non-infected nodes diffuse with diffusivity \(D_0\), infected nodes with diffusivity \(D_1\). As the number of infected nodes grows, diffusivities are additionally scaled by the factor \((1 - N_{inf} / N)^{\gamma}\).

The network evolves in descreet time steps of length dt. If left unspecified, a suitable dt is chosen automatically based on the diffusivities and contact radius.

Usage

brownian_proximity_temporalnetwork(
  size,
  avg_degree,
  radius,
  D0,
  D1 = NULL,
  gamma = 0,
  dt = NULL
)

Arguments

size

the number of nodes in the network

avg_degree

the average number of neighbours of each node

radius

the contact radius, i.e. the distance below which two nodes are connected by links

D0

the diffusivity of non-infected nodes

D1

the diffusivity of infected nodes

gamma

the exponenent which with diffusivities are scaled as the epidemic grows

dt

the time step used when evolving the network

Value

a network object

See also