Empirical temporal network comprising short contacts between nodes

Can read networks from a file, or download networks packages for NEXT-Net directly from the NEXTNet-EmpiricalNetworks repository.

Each line of the file must list a single contact in the form "" where and are integers specifying the nodes and is a floating-point value. The highest node index appearing within and defines the size of the network. Finite-duration contacts extend by dt/2 around the time specified in the file.

Usage

empirical_contact_temporalnetwork(
  path,
  dt,
  weight,
  finite_duration = FALSE,
  name = NULL,
  group = "contact",
  gzip = grepl("\\.gz$", path),
  download.timeout = 300
)

Arguments

path

name of the file

dt

duration of contacts

weight

weight of contacts, instantaneous contacts have effective weight weight * dt.

finite_duration

whether to treat contacts as having finite duration dt or to be instantaneous with effective weight weight * dt

name

name of a packaged empirical network, see packaged_empirical_network

group

packaged empirical network group containing the network

gzip

whether the file is compressed

download.timeout

for packaged networks the download timeout

See also