Computes the reproduction matrix \(M\), reproduction number \(R\), and other statistics related to these quantites.

Computes the reproduction matrix \(M\), reproduction number \(R\), and other statistics related to these quantites.

Usage

network_reproduction_matrix(nw)

Value

a named list containing entries M, r, k1, k2, k3, m_bar, R0, R_r, R_pert. See Details for a description of these quantites

Details

Computes the following quantities

M: Reproduction matrix, \(M_{ij}\) is the number of susceptible individuals of \(i\)-th smallest degree connected to a node of \(j\)-th smallest degree.

r: Degree correlation (assortativity)

k1: first raw moment of the degree distribution.

k2: second raw moment of the degree distribution.

k3: third raw moment of the degree distribution.

m1: average number of triangles a link is part of.

m2: second moment related to \(m_1\).

R0: basic reproduction number \(R_0 = (k_2 - k_1) / k_1\)

R_r: reproduction number taking assortativity but not clustering into accounts, like \(R_{pert}\) but with \(m_1\), \(m_2\) set to zero.

R_pert: reproduction number estimated from \(R_0\), \(r\), \(m_1\), \(m_2\) and \(k_1\), \(k_2\), \(k_3\).

See Cure, Pflug & Pigolotti, Exponential rate of epidemic spreading on complex networks, Physical Review E, 111, 044311 for a detailed discussion of these quantities