Creating networks

In NEXTNet, networks are directed graphs, i.e. fully defined by a set of nodes \(V\) and edges (or links) \(E \subset V \times V\) where self-edges i.e. links of the form \((v,v)\) are forbidden. NEXTNet provides various different types of algorithms for creating synthetic networks, as well as networks with are defined by specifying \(V\) and \(E\).

In addition to the plain (unweighted and static) networks defined above, NEXTNet also supports temporal and weighted networks.

In temporal networks, the set of edges/links depends on the time \(t\), i.e. \(E=E(t)\). On such networks, epidemics can only spread from a node \(u\) to a node \(v\) at times where \((u,v) \in E(t)\). At other times, the infectiousness of \(u\) for the particular link \((u,v)\) is effectively zero.

In weighted networks, each \(e \in E\) has an assigned weight \(w_e \geq 0\). These weights modulate the infectiousness of nodes, see the discussion in time_distributions. Weighted networks can be interpreted as a limit case of temporal networks in which edges fluctuate with at a very high frequency. The weight then expresses the fraction of times at which the link is present.

NEXTNetR supports the following types of static, unweighted networks:

empirical_network: Network defined by an arbitrary adjacency list read from a file.

adjacencylist_network Network defined by an arbitrary adjacency list.

erdos_renyi_network: Erdős–Rényi network, i.e. edges are sampled i.i.d from a fully-connected network.

fully_connected_network: Fully connected network, i.e. all possible edges exist.

acyclic_network: Tree-shaped network.

configmodel_network: Network with specified number of nodes of a certain degree.

configmodel_clustered_network: Configuration model with clustering.

wattsstrogatz_network:

barabasialbert_network: Barabási–Albert prefertial attachment network.

cubiclattice_network: Cubic lattice in 2 up to 8 dimensions.

the following static weighted networks

empirical_weightednetwork: Weighted network defined by an adjacency list read from a file

adjacencylist_weightednetwork: Network defined by an arbitrary adjacency list with weighted edges.

erdos_renyi_weightednetwork: Erdős–Rényi network with i.i.d edge weights.

and the following temporal networks

empirical_contact_temporalnetwork: Network defined by contacts at pre-defined times read from a file.

erdos_renyi_temporalnetwork: Erdős–Rényi network with temporally evolving edges.

brownian_proximity_temporalnetwork: Proximity network for Brownian particles in two dimensions.

sirx_temporalnetwork: Network version of the SIRX model proposed by Maier & Brockmann, 2020