Network SIRX model

This temporal network implements a network version of the SIRX model which was introduced for well-mixed populations by Maier & Brockmann. Starting from an arbitrary network, nodes are removed with baseline rate \(\kappa_0\) and infected nodes are removed with elevated rate \(\kappa_0 + \kappa\), see Maier & Brockmann 2020, Science 368 (6492), 742-74 for detais.

Usage

sirx_temporalnetwork(network, kappa0, kappa)

Arguments

network

base network. must be simple (i.e. no self-edges or multi-edges).

kappa0

\(\kappa_0\)

kappa

\(\kappa\)

See also